Fix up volume position on generation with V key

+++ TombLib/TombLib/LevelData/Room.cs
@@ -937,9 +937,9 @@ namespace TombLib.LevelData
                 NumZSectors * (0.5f * Level.WorldUnit));
         }
 
-        public Vector3 GetFloorMidpointPosition(int sectorX, int sectorZ)
+        public Vector3 GetFloorMidpointPosition(float sectorX, float sectorZ)
         {
-            var block = GetBlockTry(new VectorInt2(sectorX, sectorZ));
+            var block = GetBlockTry(new VectorInt2((int)sectorX, (int)sectorZ));
             int y = block == null ? 0 : (block.Floor.XnZp + block.Floor.XpZp + block.Floor.XpZn + block.Floor.XnZn) / 4;
             return new Vector3(sectorX * Level.WorldUnit + Level.HalfWorldUnit, y * Level.QuarterWorldUnit, sectorZ * Level.WorldUnit + Level.HalfWorldUnit);
         }
